Test for loops, then subcircuit block straddling
Circuits that begin outside of a loop may not end inside it.  The
current mechanism identifies if the block statement is a loop (by
checking if iterations > 1) before concluding a straddling is occurring.
Because let values may be present in the circuit, this test can fail for
valid circuits.

Move the test for the iterations parameter to the end, only after
determining that a block straddling is occurring.

Signed-off-by: Antonio Russo <arusso@sandia.gov>