Align simulated_probability shape with others
simulated_probability takes the same value on different subcircuit
indexes (as of now).  However, context dependent error models may change
that in the future.  Also, tooling can be kept uniform if the shape of
simulated_probability is the same as other arrays, e.g.,
normalized_counts.

Make that shape the same, but do not change the shape of
probability_by_*, which has a preexisting promise to fulfill.

Signed-off-by: Antonio Russo <arusso@sandia.gov>